requirements.txt
Django==6.0.3
pymongo==4.16.0
certifi==2026.2.25
google-api-python-client==2.179.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
